LayoutTests/imported/w3c/web-platform-tests/referrer-policy/strict-origin-when-cross-origin/meta-referrer/same-origin/http-http/iframe-tag/swap-origin-redirect/same-insecure.http-expected.txt



FAIL The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via iframe-tag using the meta-referrer
                                 delivery method with swap-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/strict-origin-when-cross-origin/meta-referrer/same-origin/http-http/iframe-tag/swap-origin-redirect/same-insecure.http.html"
